requirements.txt
Appium-Python-Client==2.6.0;python_version >= '3.0'
selenium==4.9.0
pytest==7.1.2
allure-pytest==2.9.45